updating navigation bar and adding semantic-ui
