+++ src/index.js
@@ -4,6 +4,7 @@ import { BrowserRouter as Router, Route } from "react-router-dom";
 import App from './pages/App/App';
 import * as serviceWorker from './serviceWorker';
 import './index.css';
+import 'semantic-ui-css/semantic.min.css'
 
 ReactDOM.render(
   <Router>
